chore(flutter): reorganize flutter tests
