+++ flutter/measure_flutter/test/method_channel/signal_processor_test.dart
@@ -3,8 +3,8 @@ import 'package:measure_flutter/attribute_builder.dart';
 import 'package:measure_flutter/attribute_value.dart';
 import 'package:measure_flutter/src/method_channel/signal_processor.dart';
 
-import '../noop_logger.dart';
-import '../test_method_channel.dart';
+import '../utils/noop_logger.dart';
+import '../utils/test_method_channel.dart';
 
 void main() {
   late NoopLogger logger;
