+++ flutter/measure_flutter/test/utils/noop_logger.dart
